tg_bot/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.20.0
requests==2.32.3
telebot==0.0.5
urllib3==2.2.2
